tasks/cobweb/app/requirements.txt
aiohttp
aiohttp_jinja2
pillow
ani_file
colour
